pyproject.toml
[project]
name = "werent-backend"
version = "1.0.0"
description = "WeRent Backend API - Outfit rental platform backend service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.5",
    "flask-restx>=1.3.0",
    "flask-cors>=4.0.0",
    "flask-mail>=0.10.0",
    "pydantic[email]>=2.11.7",
    "psycopg2-binary>=2.9.0",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
    "pytest-cov>=6.2.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "python-dotenv>=1.0.0",
]

production = [
    "gunicorn>=21.0.0",
    "psycopg2-binary>=2.9.0",
]
